requirements.txt
transformers>=4.45.0
qwen_vl_utils
bitsandbytes
accelerate
torch
torchvision
timm
google-genai
openai
anthropic
python-dotenv
scipy
matplotlib
pillow
numpy
sentencepiece
datasets
huggingface_hub
# Note: causal-conv1d and mamba-ssm are required for Mamba/Titans but
# must be installed manually on GPU-enabled Linux (Colab).
